Take out question's back card to another component
