+++ src/components/QuestionBackCard.module.css
@@ -1,0 +1,78 @@
+.title,
+.good-subheading {
+  composes: noselect from "lib/commons.module.css";
+}
+
+.content {
+  margin-top: 1rem;
+  display: flex;
+  flex-direction: column;
+  align-items: self-start;
+}
+
+.good-subheading {
+  text-transform: uppercase;
+  font-style: italic;
+  font-weight: bold;
+  font-family: "Inter", sans-serif;
+}
+
+.content-info {
+  margin-top: 1rem;
+  margin-bottom: 1rem;
+}
+
+.sidebar {
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  padding-bottom: 1rem;
+  min-height: 200px;
+}
+
+.sidebar div {
+  min-height: 0;
+}
+
+.sidebar-extra-images {
+  display: flex;
+  flex-wrap: nowrap;
+  justify-content: center;
+  align-self: normal;
+}
+
+.sidebar,
+.sidebar-extra-images {
+  gap: 0.5rem;
+}
+.sidebar-extra-images img,
+.sidebar img {
+  height: 100%;
+  width: 100%;
+  display: block;
+}
+
+.footer {
+  min-height: 1rem;
+  width: 1rem;
+}
+
+@media (max-width: 768px) or (orientation: portrait) {
+  .form {
+    align-items: center;
+  }
+
+  .sidebar > .sidebar-extra-images {
+    display: none;
+  }
+
+  .sidebar {
+    padding-top: 1rem;
+  }
+}
+
+@media (min-width: 768px) and (not (orientation: portrait)) {
+  .content > .sidebar-extra-images {
+    display: none;
+  }
+}
